fail when AssertionError expected but not thrown
When a user expected the production code to throw an AssertionError by
using ExpectedException (together with the #handleAssertionErrors
option) but the code under test didn't do so the test was not failing.
This was because the ExpectedException rule caught its own
AssertionError issued to signal that the expected exception was not
thrown.